requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
python-dotenv==0.19.0
pymongo==3.12.0